Broke pointers again and split the mega file into smaller bits
